requirements.txt
Flask==3.0.0
Pillow==10.0.0
reportlab==4.0.0
numpy==1.26.0
werkzeug==3.0.0
torch
torchvision
grad-cam
timm
